chore(deps): bump vst3 binding 0.1.2 -> 0.3.0
The 0.3 bindgen widened TChar from i16 to u16 (String128 is now [u16; 128]).
Adjusted the host IAttributeList string methods (setString/getString),
HostApplication::getName, and the String128 conversion accordingly; the
attribute store now holds Vec<u16>. No other API changes were required.

Verified end-to-end on macOS: Dexed and TyrellN6 selftest OK (discover,
load, params, audio), and Dexed state save/restore round-trips.

+++ vst3-host/src/internal/utils.rs
@@ -24,7 +24,7 @@ pub fn vst_string_to_string(vst_str: &String128) -> String {
         if ch == 0 {
             break;
         }
-        utf16_vec.push(ch as u16);
+        utf16_vec.push(ch);
     }
 
     String::from_utf16_lossy(&utf16_vec)
